+++ internal/controllers/reconciliation/controller.go
@@ -6,7 +6,6 @@ import (
 	"errors"
 	"fmt"
 	"os"
-	"sort"
 	"time"
 
 	"k8s.io/apimachinery/pkg/api/meta"
@@ -152,7 +151,10 @@ func (c *Controller) Reconcile(ctx context.Context, req *reconstitution.Request)
 			now := metav1.Now()
 			ready = &now
 		} else {
-			ready = evalReadinessChecks(ctx, resource, current)
+			readiness, ok := resource.ReadinessChecks.Eval(ctx, current)
+			if ok {
+				ready = &readiness.ReadyTime
+			}
 		}
 	}
 
@@ -333,34 +335,4 @@ func mungePatch(patch []byte, rv string) ([]byte, error) {
 	}
 
 	return json.Marshal(patchMap)
-}
-
-// evalReadinessChecks evaluates and prioritizes readiness checks.
-// - Nil is returned when less than all of the checks are ready
-// - If some precise and some inprecise times are given, the precise times are favored
-// - Within precise or non-precise times, the max of that group is always used
-func evalReadinessChecks(ctx context.Context, resource *reconstitution.Resource, current *unstructured.Unstructured) *metav1.Time {
-	var all []*reconstitution.Readiness
-	for _, check := range resource.ReadinessChecks {
-		if r, ok := check.Eval(ctx, current); ok {
-			all = append(all, r)
-		}
-	}
-	if len(all) == 0 || len(all) != len(resource.ReadinessChecks) {
-		return nil
-	}
-
-	sort.Slice(all, func(i, j int) bool { return all[j].ReadyTime.Before(&all[i].ReadyTime) })
-
-	// Use the max precise time if any are precise
-	for _, ready := range all {
-		ready := ready
-		if !ready.PreciseTime {
-			continue
-		}
-		return &ready.ReadyTime
-	}
-
-	// We don't have any precise times, fall back to the max
-	return &all[0].ReadyTime
 }
