+++ internal/reconstitution/api.go
@@ -5,16 +5,13 @@ import (
 	"sync"
 	"time"
 
-	celtypes "github.com/google/cel-go/common/types"
-	"github.com/google/cel-go/common/types/ref"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/types"
 	ctrl "sigs.k8s.io/controller-runtime"
 
 	apiv1 "github.com/Azure/eno/api/v1"
-	"github.com/google/cel-go/cel"
+	"github.com/Azure/eno/internal/readiness"
 )
 
 // Reconciler is implemented by types that can reconcile individual, reconstituted resources.
@@ -46,7 +43,7 @@ type Resource struct {
 	Manifest        *apiv1.Manifest
 	GVK             schema.GroupVersionKind
 	SliceDeleted    bool
-	ReadinessChecks []*ReadinessCheck
+	ReadinessChecks readiness.ReadinessChecks
 }
 
 func (r *Resource) Deleted() bool { return r.SliceDeleted || r.Manifest.Deleted }
@@ -54,81 +51,6 @@ func (r *Resource) Deleted() bool { return r.SliceDeleted || r.Manifest.Deleted 
 func (r *Resource) Parse() (*unstructured.Unstructured, error) {
 	u := &unstructured.Unstructured{}
 	return u, u.UnmarshalJSON([]byte(r.Manifest.Manifest))
-}
-
-type ReadinessCheck struct {
-	Name    string
-	program cel.Program
-	env     *cel.Env
-}
-
-func MustReadinessCheckTest(expr string) *ReadinessCheck {
-	env, err := newCelEnv()
-	if err != nil {
-		panic(err)
-	}
-	check, err := newReadinessCheck(env, expr)
-	if err != nil {
-		panic(err)
-	}
-	return check
-}
-
-func newReadinessCheck(env *cel.Env, expr string) (*ReadinessCheck, error) {
-	ast, iss := env.Compile(expr)
-	if iss != nil && iss.Err() != nil {
-		return nil, iss.Err()
-	}
-	prgm, err := env.Program(ast) // TODO: Set InterruptCheckFrequency
-	if err != nil {
-		return nil, err
-	}
-	return &ReadinessCheck{program: prgm, env: env}, nil
-}
-
-func (r *ReadinessCheck) Eval(ctx context.Context, resource *unstructured.Unstructured) (*Readiness, bool) {
-	if resource == nil {
-		return nil, false
-	}
-	val, details, err := r.program.ContextEval(ctx, map[string]any{"self": resource.Object})
-	if details != nil {
-		cost := details.ActualCost()
-		if cost != nil {
-			celEvalCost.Add(float64(*cost))
-		}
-	}
-	if err != nil {
-		return nil, false
-	}
-
-	// Support matching on condition structs.
-	// This allows us to grab the transition time instead of just using the current time.
-	if list, ok := val.Value().([]ref.Val); ok {
-		for _, ref := range list {
-			if mp, ok := ref.Value().(map[string]any); ok {
-				if mp != nil && mp["status"] == "True" && mp["type"] != nil && mp["reason"] != nil {
-					ts := metav1.Now()
-					if str, ok := mp["lastTransitionTime"].(string); ok {
-						parsed, err := time.Parse(time.RFC3339, str)
-						if err == nil {
-							ts.Time = parsed
-						}
-					}
-					return &Readiness{ReadyTime: ts, PreciseTime: err == nil}, true
-				}
-			}
-		}
-	}
-
-	if val == celtypes.True {
-		return &Readiness{ReadyTime: metav1.Now()}, true
-	}
-	return nil, false
-}
-
-type Readiness struct {
-	ReadyTime   metav1.Time
-	PreciseTime bool // true when time came from a condition, not the controller's metav1.Now
 }
 
 // ResourceRef refers to a specific synthesized resource.
